Move getcurenttime in Utils
