--- ToolBox/Dockerfile
@@ -1,16 +1,0 @@
-FROM python:3.10
-
-WORKDIR /app
-
-RUN pip install poetry
-
-# Install dependencies
-COPY poetry.lock pyproject.toml /app/
-RUN poetry config virtualenvs.create false && poetry install --no-dev
-
-COPY . .
-
-RUN useradd -ms /bin/bash sekoiaio-runtime
-USER sekoiaio-runtime
-
-ENTRYPOINT [ "python", "./main.py" ]
