--- ToolBox/getcurrenttime_modules/__init__.py
@@ -1,6 +1,0 @@
-from sekoia_automation.module import Module
-from getcurrenttime_modules.models import GetcurrenttimeModuleConfiguration
-
-
-class GetcurrenttimeModule(Module):
-    configuration: GetcurrenttimeModuleConfiguration
