--- ToolBox/getcurrenttime_modules/models.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel, Field
-
-
-class GetcurrenttimeModuleConfiguration(BaseModel):
-    pass
