--- ToolBox/manifest.json
@@ -1,12 +1,0 @@
-{
-  "configuration": {
-    "title": "GetcurrenttimeModuleConfiguration",
-    "type": "object",
-    "properties": {}
-  },
-  "description": "Get the current time of a alerte depend of UTC",
-  "name": "GetCurrentTime",
-  "uuid": "c42eef31-82b7-4c25-aef2-1417332a380c",
-  "docker": "sekoia-automation-module-getcurrenttime",
-  "version": "1.0"
-}
