--- ToolBox/tests/conftest.py
@@ -1,17 +1,0 @@
-from shutil import rmtree
-from tempfile import mkdtemp
-
-import pytest
-
-from sekoia_automation import constants
-
-
-@pytest.fixture
-def symphony_storage():
-    original_storage = constants.DATA_STORAGE
-    constants.DATA_STORAGE = mkdtemp()
-
-    yield constants.DATA_STORAGE
-
-    rmtree(constants.DATA_STORAGE)
-    constants.SYMPHONY_STORAGE = original_storage
